+++ CHANGES.md
@@ -6,6 +6,8 @@
  - no default physics any more must be
    useAmmo, useJolt, useCannon or useMatter
  - Improved volumetric and light - range param
+ - PlaneCamera added - with zoom , support for mobile 
+ 
 
 [1.14.0]
  - Added HZB postproccesing effect